requirements.txt
PyYAML>=6.0,<7.0
yt-dlp>=2026.3.3,<2027.0
curl-cffi>=0.7,<1.0
